This is The Initial Commit
I've Started on the main.c file, where the main Flight-code will live.

I've outlined all of the System Clock initialization files and the code to start the ports for the various communication protocols: USART, SPI, the ADC, etc.

These just have dummy ports assigned to them, and we will need to talk with Electrical to figure out which ones we are actually using.

I also filled out a bunch of pseudo-code in the form of comments.

I outlined the reset protocols and the basic flight-state logic.

I tried to set up the libraries for the editing environment, but I might have missed importing a few. If you get any weird erros later on, this is probably why. That or we forgot to update a port correctly (or initialize a port, LOL).