Started Pressor Sensor Drivers and SPI drivers
Trying to make all drivers as port independant as possible. That means, allowing the SPI initialization to accept a port specification. It should return some sort of flag indicating if that port has valid SPI on it for this board.

I need to look in the data sheet to figure out which ports have SPI and all that stuff, and that's a bore.