Started Prototype Board
Working on hooking up the BNO055 right now.

It's apparently easier to set it up on I2C, so I'm trying that.

The library wasn't working initially but the problem was that tht ASF header didn't have everything defined that it needed through the ASF Wizard interface.

So, yeah. Remember that if you get weird errors to check the ASF Wizard first to make sure you have everything installed properly in the standard library.

I also found some libraries for UART and I2C/TWI that make it much easier to deal with things hooked up to those interfaces. I've installed them in the prototype and placed their files in the drive.